Move NumericRank enum to constants
